Improve and move filters to localfile-config
